added random palette

+++ graphics/palette/gradient_palette.cpp
@@ -7,7 +7,7 @@ gradient_palette::gradient_palette(double min, double max, color cmin, color cma
     this->cmax = cmax;
 }
 
-color gradient_palette::compute_color(double value) {
+color gradient_palette::compute_color(int value) {
     unsigned char R = compute_component(value, cmin.R, cmax.R);
     unsigned char G = compute_component(value, cmin.G, cmax.G);
     unsigned char B = compute_component(value, cmin.B, cmax.B);
@@ -19,7 +19,7 @@ gradient_palette::gradient_palette() {
     cmax = color(255,255,255);
 }
 
-unsigned char gradient_palette::compute_component(double value, unsigned char v1, unsigned char v2) {
+unsigned char gradient_palette::compute_component(int value, unsigned char v1, unsigned char v2) {
     double normalized = (value - min)/(max - min);
     return v1 + (unsigned char)(normalized * (v2-v1));
 }
